requirement.txt
roma
trimesh
drake
open_clip-torch
sentence-transformers
matplotlib
numpy
pyyaml
compress_pickle
accelerate